+++ data/scripting/catalog.lua
@@ -2,6 +2,7 @@ local raw = trxc.catalog
 
 local catalog = {
   objects = raw.objects,
+  flip_effects = raw.flip_effects,
 }
 
 trx.catalog = catalog