+++ docs/CHANGELOG.md
@@ -4,6 +4,9 @@
 - added the ability to disable manual camera (Gameplay → Controls → Manual camera)
 - added the ability to enable bouncy grenades (Gameplay → General → Enable bouncy grenades)
 - added a new Lua event, `trx.events.on_game_start`, which fires when the level finishes loading and the game is about to start
+- added a new Lua function, `trx.rooms.flip()`, to toggle the flip map (#4704)
+- added a new Lua function, `trx.rooms.flip_effect()`, to set the active flip effect with an optional timer (#4704)
+- added a new Lua catalog, `trx.catalog.flip_effects` for name-based flip effect catalog IDs
 - added support for cross-fades to the title screen
 - added visual previews of bar colors (Graphic Options → Bars)
 - added endless sprint (available previously via the `/restless` command) to the UI settings (Gameplay options → Mods → Endless sprint)